+++ CSE/AgreeEscort.cpp
@@ -66,11 +66,11 @@ void AgreeEscort::EmotionalReaction(NPC_Actor* affectingActor)
 	}
 }
 
-float AgreeEscort::NPC_CalculateInclination()
+double AgreeEscort::NPC_CalculateInclination()
 {
 	double a = static_cast<NPC_Actor*>(subject)->Get_Conscientious();
 	double b = static_cast<NPC_Actor*>(subject)->Get_pAgreeable(object->GetID());
 	double w = static_cast<NPC_Actor*>(subject)->Get_Agreeable();
 	double result = Blend(a, b, w) - 0.5f /*for travel*/;
-	return float(result);
+	return result;
 }
