+++ CSE/Apologize.cpp
@@ -60,11 +60,11 @@ void Apologize::EmotionalReaction(NPC_Actor* affectingActor)
 	}
 }
 
-float Apologize::NPC_CalculateInclination()
+double Apologize::NPC_CalculateInclination()
 {
 	double a = static_cast<NPC_Actor*>(subject)->Get_Agreeable();
 	double b = static_cast<NPC_Actor*>(subject)->Get_pAgreeable(object->GetID());
 	double w = static_cast<NPC_Actor*>(subject)->Get_Happy();
 	double result = Blend(a, b, w);
-	return float(result);
+	return result;
 }
