+++ CSE/Arrive.h
@@ -21,7 +21,7 @@ public:
 
 	virtual void ExecuteConsequences(WorldState*);
 	virtual void EmotionalReaction(NPC_Actor*);
-	virtual float NPC_CalculateInclination();
+	virtual double NPC_CalculateInclination();
 
 	virtual bool ReadyToExecute();
 };
