+++ CSE/Greet.cpp
@@ -85,11 +85,11 @@ void Greet::EmotionalReaction(NPC_Actor* affectingActor)
 	}
 }
 
-float Greet::NPC_CalculateInclination()
+double Greet::NPC_CalculateInclination()
 {
 	double a = static_cast<NPC_Actor*>(subject)->Get_Agreeable();
 	double b = static_cast<NPC_Actor*>(subject)->Get_pAgreeable(object->GetID());
 	double w = static_cast<NPC_Actor*>(subject)->Get_Happy();
 	double result = Blend(a, b, w);
-	return float(result);
+	return result;
 }
