+++ CSE/Hug.cpp
@@ -81,7 +81,7 @@ void Hug::EmotionalReaction(NPC_Actor* affectingActor)
 	}
 }
 
-float Hug::NPC_CalculateInclination()
+double Hug::NPC_CalculateInclination()
 {
 	double a = static_cast<NPC_Actor*>(subject)->Get_pAgreeable(object->GetID());
 	double b = static_cast<NPC_Actor*>(subject)->Get_Extraverted();
