+++ CSE/Hug.h
@@ -13,5 +13,5 @@ public:
 	std::string GetSentence();
 	void ExecuteConsequences(WorldState*);
 	void EmotionalReaction(NPC_Actor*);
-	float NPC_CalculateInclination();
+	double NPC_CalculateInclination();
 };
