+++ CSE/Insult.h
@@ -12,6 +12,6 @@ public:
 
 	void ExecuteConsequences(WorldState*);
 	void EmotionalReaction(NPC_Actor*);
-	float NPC_CalculateInclination();
+	double NPC_CalculateInclination();
 };
