+++ CSE/Reprimand.h
@@ -17,6 +17,6 @@ public:
 
 	void ExecuteConsequences(WorldState*);
 	void EmotionalReaction(NPC_Actor*);
-	float NPC_CalculateInclination();
+	double NPC_CalculateInclination();
 };
