+++ CSE/RequestEntry.h
@@ -16,6 +16,6 @@ public:
 
 	virtual void ExecuteConsequences(WorldState*);
 	virtual void EmotionalReaction(NPC_Actor*);
-	virtual float NPC_CalculateInclination();
+	virtual double NPC_CalculateInclination();
 };
