+++ CSE/SuggestFlowers.h
@@ -18,6 +18,6 @@ public:
 
 	void ExecuteConsequences(WorldState*);
 	void EmotionalReaction(NPC_Actor*);
-	float NPC_CalculateInclination();
+	double NPC_CalculateInclination();
 };
